database: Move rest of the functions to operator.
