+++ core/src/database/mod.rs
@@ -9,7 +9,6 @@
 use crate::{config::BridgeConfig, errors::BridgeError};
 use sqlx::{Pool, Postgres};
 
-mod common;
 mod header_chain_prover;
 mod operator;
 mod verifier;
